Added VPN Functionality to API. Also modified user listing to remove UIDs.

+++ bin/cluster.rb
@@ -34,9 +34,10 @@ $etcConfigPath = config['vpn']['vpnconfigpath']
 def clientMenu()  
   sel = $prompt.select('Choose an option') do |menu|
     menu.choice 'View VPN Status', 'viewClients'
-    menu.choice 'Assign VPN', 'assignSlot'
+    menu.choice 'Assign VPN', 'promptAssignSlot'
     menu.choice 'Modify VPN', 'editClientAssignment'
     menu.choice 'View Assigned VPN Client Script', 'printVPNClientScript'
+    menu.choice 'Generate Password for Client', 'promptGenClientPasswd'
     menu.choice 'Deconfigure VPN', 'deconfigureClient'
     menu.choice 'Return', 'getout'
   end
@@ -50,12 +51,14 @@ def client()
     viewVPNs()
   when 'printVPNClientScript'
     printVPNClientScript()
-  when 'assignSlot'
-    assignSlot()
+  when 'promptAssignSlot'
+    promptAssignSlot()
   when 'editClientAssignment'
     editClientAssignment()
   when 'deconfigureClient'
     promptDeconfigureVPNClientAssignment()
+  when 'promptGenClientPasswd'
+    promptGenClientPasswd()
   when 'getout'
     main()
   end
@@ -66,21 +69,53 @@ def reloadSlotFileConfig()
 end
 
 def promptSelectUsedSlot()
-  vpnSlot = $prompt.select("Choose a VPN Slot: ", slotListUsed.keys.to_a)
+  begin
+    vpnSlot = $prompt.select("Choose a VPN Slot: ", slotListUsed.keys.to_a)
+  rescue
+    return nil
+  end
   return vpnSlot
 end 
+
+def promptSelectClient()
+  vpn = []
+  slotListUsed.select do |k,v|
+    vpn << k + ' ' + v['clientname']
+  end
+  choices = []
+  vpn.each_with_index do |e,i|
+     choices << {:name => e, :value => i}
+  end
+  vpnSlot = $prompt.enum_select("Choose a VPN Slot: ", vpn)  
+end
 
 def promptAssignSlot()
   vpnSlot = $prompt.select("Choose a VPN Slot to assign: ", slotListAvail.keys.to_a)
   clientName = $prompt.ask("Client Name: ", required: true)
   assignedVPNSlot = [clientName,vpnSlot]
-  return assignedVPNSlot
+  begin
+    if assignSlot(clientName,vpnSlot)
+      puts "Successfully Assigned Slot"
+      return true
+    else
+      puts "Could not assign slot."
+      client()
+    end
+  rescue
+    puts "Could not assign slot."
+    client()
+  end
 end
 
 def promptSelectSlot()
-  vpnSlot = $prompt.select("Choose a VPN Slot to Modify: ", slotListUsed.keys.to_a)
-  return vpnSlot
+  begin
+    vpnSlot = $prompt.select("Choose a VPN Slot to Modify: ", slotListUsed.keys.to_a)
+    return vpnSlot
+  rescue
+    return nil
+  end
 end 
+
 
 def vpnSlotFileWrite(slotHash)
   $vpnSlotFileConfig['slots'].merge!(slotHash)
@@ -99,26 +134,25 @@ def slotListUsed()
   slotListUsed = $vpnSlotFileConfig['slots'].select {|k,h| h['configured']} 
 end
 
-def assignSlot()
-  clientName,vpnSlot = promptAssignSlot()
-  slotHash = $vpnSlotFileConfig ['slots'].select {|k,h| k == vpnSlot} 
-  File.open("/etc/openvpn/ccd-cluster/#{clientName}", 'w') do |l|
-    l.puts("ifconfig-push #{slotHash[vpnSlot]['clientip']} #{slotHash[vpnSlot]['serverip']}")
+def assignSlot(clientName,vpnSlot)
+  slotHash = $vpnSlotFileConfig ['slots'].select {|k,h| k == vpnSlot}
+  if ! slotHash.empty? && slotListAvail.keys.include?(vpnSlot) 
+    File.open("/etc/openvpn/ccd-cluster/#{clientName}", 'w') do |l|
+      l.puts("ifconfig-push #{slotHash[vpnSlot]['clientip']} #{slotHash[vpnSlot]['serverip']}")
+    end
+    if ! Open3.capture3("\/sbin\/useradd -M -N -s \/sbin\/nologin #{clientName} --gid vpn")[2].success?
+      return nil
+    else
+      slotHash[vpnSlot].store('clientscript', generateTemplate(clientName))
+      slotHash[vpnSlot]['clientname'] = clientName
+      slotHash[vpnSlot]['configured'] = true
+      slotHash[vpnSlot]['password'] = '' 
+      vpnSlotFileWrite(slotHash)
+      reloadSlotFileConfig()
+    end
+  else
+    return nil
   end
-  if ! Open3.capture3("\/sbin\/useradd -M -N -s \/sbin\/nologin #{clientName} --gid vpn")[2].success?
-    puts "Could not create user"
-    client()
-  end
-  puts userCreateStatus
-  rawPasswd = checkPasswd()
-  cryptedPasswd = genPasswd(rawPasswd)
-  setPasswd(clientName,cryptedPasswd)
-  slotHash[vpnSlot].store('clientscript', generateTemplate(clientName))
-  slotHash[vpnSlot]['clientname'] = clientName
-  slotHash[vpnSlot]['configured'] = true
-  slotHash[vpnSlot]['password'] = cryptedPasswd
-  vpnSlotFileWrite(slotHash)
-  reloadSlotFileConfig()
 end
 
 def setPasswd(userName,cryptedPasswd)
@@ -131,32 +165,95 @@ def setPasswd(userName,cryptedPasswd)
   end
 end
 
-def genPasswd(rawPasswd)
-  if rawPasswd
-    passwdOut,err,status = Open3.capture3("printf '#{rawPasswd}' | \/bin\/openssl passwd -1 -stdin")
-    if status.success?
-      return passwdOut[0..-2]
-    else
-      raise StandardError
-    end
+def genPasswd()
+  passwdAry = []
+  o = [('a'..'z'), ('A'..'Z')].map(&:to_a).flatten
+  generatedPasswd = (0...15).map { o[rand(o.length)] }.join
+  passwdAry = []
+  passwdAry[0] = generatedPasswd
+  passwdOut,err,status = Open3.capture3("printf '#{generatedPasswd}' | \/bin\/openssl passwd -1 -stdin")
+  if status.success?
+    passwdAry[1] = passwdOut[0..-2]
+    return passwdAry
   else
     raise StandardError
   end
 end
 
-def promptPasswd()
-  rawPasswd1 = $prompt.mask("Enter a password: ")
-  rawPasswd2 = $prompt.mask("Confirm the password: ")
-  rawPasswds = [rawPasswd1,rawPasswd2]
+def promptGenClientPasswd()
+  slot = promptSelectUsedSlot()
+  if ! slot.nil?
+    yn = $prompt.yes?("Are you sure you want to re-generate the password for #{slot}?") do |q|
+      q.default false
+      q.positive "Y"
+      q.negative "N"
+    end 
+    if yn
+      clientName = $vpnSlotFileConfig['slots'][slot]['clientname']
+      begin
+        rawPass = genClientPasswd(slot)
+      rescue
+        puts "Could not set Password."
+      end
+      if ! rawPass.nil?
+        puts ""
+        puts "One time password for #{slot} - #{clientName}: "
+        puts ""
+        puts "#{rawPass}"
+      else
+        puts "Could not set Password."
+      end
+    end
+  end
 end
 
-def checkPasswd()
-  rawPasswds = promptPasswd()
-  while rawPasswds[0] != rawPasswds[1]
-    puts "Passwords do not match - try again"
-    rawPasswds = promptPasswd() 
+
+def genClientPasswd(vpnSlot)
+  begin
+    passwdAry = genPasswd()
+    randomRawPass = passwdAry[0]
+    cryptedPasswd = passwdAry[1]
+    slotHash = $vpnSlotFileConfig ['slots'].select {|k,h| k == vpnSlot}
+    if ! slotHash.empty? || ! slotHash[vpnSlot]['configured'] == true
+      clientName = slotHash[vpnSlot]['clientname']
+      if ! setPasswd(clientName, cryptedPasswd)
+        fail
+      end
+      slotHash[vpnSlot]['password'] = cryptedPasswd 
+      vpnSlotFileWrite(slotHash)
+      reloadSlotFileConfig()
+      return randomRawPass
+    end
+  rescue
+    return nil
   end
-  return rawPasswds[0] 
+end
+
+# def promptPasswd()
+#   rawPasswd1 = $prompt.mask("Enter a password: ")
+#   rawPasswd2 = $prompt.mask("Confirm the password: ")
+#   rawPasswds = [rawPasswd1,rawPasswd2]
+# end
+
+# def checkPasswd()
+#   rawPasswds = promptPasswd()
+#   while rawPasswds[0] != rawPasswds[1]
+#     puts "Passwords do not match - try again"
+#     rawPasswds = promptPasswd() 
+#   end
+#   return rawPasswds[0] 
+# end
+
+def getVpnList()
+  slots = {}
+  slothash = {}
+  slotListUsed.each_key do |slot|
+    clientname = $vpnSlotFileConfig['slots'][slot]['clientname']
+    pingable = checkClientResponding($vpnSlotFileConfig['slots'][slot]['clientip'])
+    slothash = {:clientname => clientname, :connstatus => pingable}
+    slots.merge! slot => slothash 
+  end
+  return slots
 end
 
 def viewVPNs()
@@ -184,10 +281,32 @@ def generateTemplate(clientName)
   renderer.result()
 end
 
+def getVPNClientScript(vpnSlot)
+  begin
+    slotHash = $vpnSlotFileConfig ['slots'].select {|k,h| k == vpnSlot}
+    if ! slotHash[vpnSlot]['clientscript'].nil? 
+      return slotHash[vpnSlot]['clientscript']
+    else 
+      return nil
+    end
+  rescue
+    return nil
+  end
+end
+
+
 def printVPNClientScript()
   vpnSlot = promptSelectUsedSlot()
-  slotHash = $vpnSlotFileConfig ['slots'].select {|k,h| k == vpnSlot} 
-  puts slotHash[vpnSlot]['clientscript']
+  if ! vpnSlot.nil?
+    script = getVPNClientScript(vpnSlot)
+    if ! script.nil?
+      return script
+    else
+      puts "No VPNs assigned"
+    end
+  else
+    puts "No VPNs assigned"
+  end
 end
 
 def deleteVPNClientUser(clientUser)
@@ -195,15 +314,20 @@ def deleteVPNClientUser(clientUser)
 end
 
 def deconfigureVPNClient(deconfVPN)
-  slotHash = $vpnSlotFileConfig ['slots'].select {|k,h| k == deconfVPN} 
-  oldUser = slotHash[deconfVPN]['clientname'] 
-  slotHash[deconfVPN]['configured'] = false
-  slotHash[deconfVPN]['clientname'] = nil
-  slotHash[deconfVPN]['password'] = nil
-  slotHash[deconfVPN]['clientscript'] = nil
-  deleteVPNClientUser(oldUser)
-  vpnSlotFileWrite(slotHash)
-  reloadSlotFileConfig()
+  begin
+    slotHash = $vpnSlotFileConfig ['slots'].select {|k,h| k == deconfVPN} 
+    oldUser = slotHash[deconfVPN]['clientname'] 
+    slotHash[deconfVPN]['configured'] = false
+    slotHash[deconfVPN]['clientname'] = nil
+    slotHash[deconfVPN]['password'] = nil
+    slotHash[deconfVPN]['clientscript'] = nil
+    deleteVPNClientUser(oldUser)
+    vpnSlotFileWrite(slotHash)
+    reloadSlotFileConfig()
+    return true
+  rescue
+    return nil
+  end
 end
 
 def promptDeconfigureVPNClientAssignment()
@@ -212,33 +336,41 @@ def promptDeconfigureVPNClientAssignment()
     clientMenu()
   else
     deconfVPN = $prompt.select("Choose a VPN Slot to Deconfigure: ", slotListUsed.keys.to_a)
-    deconfigureVPNClient(deconfVPN)
+    if ! deconfigureVPNClient(deconfVPN)
+      puts "Deconfigure Failed."
+    else
+      puts "Successfully Deconfigured #{deconfVPN}"
+    end
   end
 end
 
 def editClientAssignment()
   vpnSlot = promptSelectSlot()
-  slotHash = $vpnSlotFileConfig ['slots'].select {|k,h| k == vpnSlot} 
-  oldName = slotHash[vpnSlot]['clientname']
-  newName = $prompt.ask("Client Name: ", required: true)
-  if $vpnSlotFileConfig['slots'].each
-    puts "Already a Client with that name"
-    return false
+  if ! vpnSlot.nil?
+    slotHash = $vpnSlotFileConfig ['slots'].select {|k,h| k == vpnSlot} 
+    oldName = slotHash[vpnSlot]['clientname']
+    newName = $prompt.ask("Client Name: ", required: true)
+    if $vpnSlotFileConfig['slots'].each
+      puts "Already a Client with that name"
+      return false
+    else
+      deleteVPNClientUserStatus = deleteVPNClientUser(oldName)
+      if ! deleteVPNClientUserStatus[2].success? 
+        puts deleteVPNClientUserStatus[1] + ' ' + deleteVPNClientUserStatus[0] 
+      end
+      if File.exist?("/etc/openvpn/ccd-cluster/#{oldName}") 
+        File.delete("/etc/openvpn/ccd-cluster/#{oldName}") 
+      end
+      File.open("/etc/openvpn/ccd-cluster/#{newName}", 'w') do |l|
+        l.puts("ifconfig-push #{slotHash[vpnSlot]['clientip']} #{slotHash[vpnSlot]['serverip']}")
+      end
+      slotHash[vpnSlot]['clientname'] = newName
+      vpnSlotFileWrite(slotHash)
+      reloadSlotFileConfig()
+    end 
   else
-    deleteVPNClientUserStatus = deleteVPNClientUser(oldName)
-    if ! deleteVPNClientUserStatus[2].success? 
-      puts deleteVPNClientUserStatus[1] + ' ' + deleteVPNClientUserStatus[0] 
-    end
-    if File.exist?("/etc/openvpn/ccd-cluster/#{oldName}") 
-      File.delete("/etc/openvpn/ccd-cluster/#{oldName}") 
-    end
-    File.open("/etc/openvpn/ccd-cluster/#{newName}", 'w') do |l|
-      l.puts("ifconfig-push #{slotHash[vpnSlot]['clientip']} #{slotHash[vpnSlot]['serverip']}")
-    end
-    slotHash[vpnSlot]['clientname'] = newName
-    vpnSlotFileWrite(slotHash)
-    reloadSlotFileConfig()
-  end 
+    puts "No configured VPNs"
+  end
 end
 
 def checkClientResponding(vpnSlotClientIP)
